knowhere build ok

+++ core/src/wrapper/VecImpl.h
@@ -70,7 +70,7 @@ class VecIndexImpl : public VecIndex {
     Search(const int64_t& nq, const float* xq, float* dist, int64_t* ids, const Config& cfg) override;
 
     Status
-    SearchById(const int64_t& nq, const float* xq, faiss::ConcurrentBitsetPtr bitset, float* dist, int64_t* ids,
+    SearchById(const int64_t& nq, const int64_t* xq, faiss::ConcurrentBitsetPtr bitset, float* dist, int64_t* ids,
                const Config& cfg) override;
 
     Status
